### 2022-07-26 #4612 @editorialbot
[PRE REVIEW]: graphenv: a Python library for reinforcement learning on graph search spaces
**Submitting author:** @pstjohn (<a href="http://orcid.org/0000-0002-7928-3722">Peter C. St. John</a>)
**Repository:** https://github.com/NREL/graph-env/
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.5
