### 2022-07-27 #4612 @osorensen
@pstjohn, thanks for submitting this article to JOSS. I'll try to find reviewers now.

In the meantime, could you please correct the following reference to Zhou (2019) on line 17. Seems like the citation key is wrong, because it does not appear in the bibliography at the end either.
