### 2022-07-27 #4612 @pstjohn
Thanks @osorensen for handling this!

I fixed those issues with the manuscript -- We missed the '@' symbol for Zhou_2019
