### 2022-07-27 9bea3ed
Update paper.md
switching ORCiD IDs to correct names
