### 2022-07-29 #4612 @idoby
> 👋 @idoby @Viech would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Would love to help. This would be my first review so bear with me please.
