### 2022-07-29 #4621 @editorialbot
[REVIEW]: graphenv: a Python library for reinforcement learning on graph search spaces
**Submitting author:** @pstjohn (<a href="http://orcid.org/0000-0002-7928-3722">Peter C. St. John</a>)
**Repository:** https://github.com/NREL/graph-env/
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.3

@iammix & @vwxyzjn & @Viech & @idoby, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
