### 2022-07-29 #4621 @idoby
This seems like a very interesting package! I will have the time to dive in more in-depth soon, but in the meanwhile, can the authors fix the citation on line 17 of the paper to be consistent with the other citations, and resolve the invalid citation issue?

Also, please clarify the contributions of the two extra authors who are not listed as collaborators on the GitHub repo.

Thanks!
