+++ joss/paper.bib
@@ -61,7 +61,7 @@
 @Article{	  Sowndarya_S_V_2021,
   title		= {A quantitative metric for organic radical stability and
 		  persistence using thermodynamic and kinetic features},
-  author	= {Sowndarya, Shree S. V. and  {St. John}, Peter C. and Paton, Robert S.},
+  author	= {S.V., Shree S. and  {St. John}, Peter C. and Paton, Robert S.},
   year		= 2021,
   journal	= {Chemical Science},
   publisher	= {Royal Society of Chemistry ({RSC})},
@@ -103,7 +103,7 @@
 
 @Article{	  sv2021multi,
   title		= {Multi-objective goal-directed optimization of de novo stable organic radicals for aqueous redox flow batteries},
-  author	= {Sowndarya, Shree S. V. and Law, Jeffrey and Tripp, Charles
+  author	= {S.V., Shree S. and Law, Jeffrey and Tripp, Charles
 		  and Duplyakin, Dmitry and Skordilis, Erotokritos and
 		  Biagioni, David and Paton, Robert and {St. John}, Peter},
   year		= 2022,
