### 2022-07-30 #4621 @idoby
> Apologies if these contribution edits should have been made in the manuscript?
>
> * Dmitry Duplyakin contributed documentation for running graph-env on AWS, and debugged a lot of that workflow  ([Add config and doc for TSP on Ray at AWS NREL/graph-env#22](https://github.com/NREL/graph-env/pull/22)).
> * Jeffrey Law has contributed extensively to downstream repositories [NREL/rlmolecule](https://github.com/nrel/rlmolecule), and his feedback has been helpful in the direction of this more general-purpose library

I'm pretty new at reviewing for JOSS, but the documentation does not state this information has to be specified in the manuscript. So I think just your comment should be fine.

> That invalid DOI, 10.1038/s42256-022-00506-3, should be live starting 04 August 2022 at 11:00 (US Eastern Time)

OK, LGTM then. Appreciate your work on the library and thank you for contributing it. :)
