### 2022-08-08 NREL/graph-env#33 @vwxyzjn
Adopt a package manager
Hi @pstjohn, this is a cool project!

Have you considered adopting a package manager like [poetry](https://python-poetry.org/)? The current way of installing with `setup.py` is deprecated by PyPa (see https://pip.pypa.io/en/stable/reference/build-system/setup-py/#setup-py-legacy). For future projects, PyPa recommends using `pyproject.toml`, which can be automatically generated by `poetry`. Poetry also has many benefits, like locking the dependencies to ensure reproducibility — an important consideration for research projects.

Feel free to checkout how we use poetry in CleanRL ([docs](https://docs.cleanrl.dev/get-started/installation/) and [pyproject.toml](https://github.com/vwxyzjn/cleanrl/blob/master/pyproject.toml)).
