### 2022-08-08 NREL/graph-env#34 @vwxyzjn
Make the NN packages optional
Please consider making the neural network / rl libraries optional in the list of dependencies. Researchers should be able to freely choose their framework of choice without having to install RLLib or tensorflow.

One possible resolution is to use the optional dependency specs in poetry (see #33 and [here](https://github.com/vwxyzjn/cleanrl/blob/1878a4f5e29e848135c6e08f89afdaa744403b48/pyproject.toml#L54-L57) as an example), so users can do things like `poetry install -E my-rl-optional-packages`
