### 2022-08-08 #4621 @vwxyzjn
Opened a couple of issues:

https://github.com/NREL/graph-env/issues/33
https://github.com/NREL/graph-env/issues/34

Also, please consider filling out the github repo's description and keywords. See the screenshot below as an example.
