### 2022-08-09 3233832
initial tests with pyproject.toml and pip-based dependency installs
