+++ setup.cfg
@@ -12,17 +12,17 @@ classifiers =
 packages = find:
 install_requires =
   gym
-  tensorflow
   networkx
   ray[tune,rllib]
-  torch
 
-[versioneer]
-VCS = git
-style = pep440
-versionfile_source = graphenv/_version.py
-versionfile_build = graphenv/_version.py
-tag_prefix =
+[options.extras_require]
+tensorflow = tensorflow
+torch = torch
+dev =
+  torch
+  tensorflow
+  pytest
+  build
 
 [flake8]
 max-line-length = 88
