+++ .github/workflows/run_tests_and_deploy.yml
@@ -18,11 +18,11 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [ 3.7, 3.8, 3.9 ]
+        python-version: ["3.7", "3.8", "3.9", "3.10"]
 
     steps:
       -
-        uses: actions/checkout@v2
+        uses: actions/checkout@v3
       -
         name: Additional info about the build
         shell: bash
@@ -31,24 +31,19 @@ jobs:
           df -h
           ulimit -a
       - 
-        uses: conda-incubator/setup-miniconda@v2
+        name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
-          environment-file: devtools/conda-envs/environment.yml
-          channels: conda-forge,defaults
-          activate-environment: test
-          auto-update-conda: false
-          auto-activate-base: false
-          show-channel-urls: true
       -
         name: Install package
         shell: bash -l {0}
         run: |
-          python -m pip install . --no-deps
-          conda list
+          python -m pip install --upgrade pip
+          python -m pip install .[dev]
       -
         name: Run tests
-        shell: bash -l {0}        
+        shell: bash -l {0}
         run: pytest -v tests/
 
   release:
@@ -70,7 +65,13 @@ jobs:
           python -m pip install --upgrade pip setuptools wheel    
       - 
         name: Build package
-        run: python setup.py sdist bdist_wheel    
+        run: >-
+          python -m
+          build
+          --sdist
+          --wheel
+          --outdir dist/
+          .
       - 
         name: Publish package
         uses: pypa/gh-action-pypi-publish@release/v1
