+++ graphenv/__init__.py
@@ -1,6 +1,6 @@
 from ray.rllib.utils.framework import try_import_tf, try_import_torch
 
-from _version import __version__, __version_tuple__  # noqa: F401
+from graphenv._version import __version__, __version_tuple__  # noqa: F401
 
 tf1, tf, tfv = try_import_tf()
 assert tfv == 2
