+++ graphenv/__init__.py
@@ -1,6 +1,10 @@
 from ray.rllib.utils.framework import try_import_tf, try_import_torch
 
-from graphenv._version import __version__, __version_tuple__  # noqa: F401
+try:
+    from graphenv._version import __version__, __version_tuple__  # noqa: F401
+except ModuleNotFoundError:
+    __version__ = "package not installed"
+    __version_tuple__ = (None, None, None)
 
 tf1, tf, tfv = try_import_tf()
 assert tfv == 2
