### 2022-08-09 9250107
remove nfp tests
