### 2022-08-09 4750224
ensure setuptools_scm is installed
