+++ setup.cfg
@@ -23,6 +23,7 @@ dev =
   tensorflow
   pytest
   build
+  setuptools_scm
 
 [flake8]
 max-line-length = 88
