+++ .github/workflows/run_tests_and_deploy.yml
@@ -62,7 +62,7 @@ jobs:
       - 
         name: Install dependencies
         run: |
-          python -m pip install --upgrade pip setuptools wheel    
+          python -m pip install --upgrade pip setuptools wheel setuptools_scm
       - 
         name: Build package
         run: >-
