+++ .github/workflows/run_tests_and_deploy.yml
@@ -39,7 +39,8 @@ jobs:
         name: Install package
         shell: bash -l {0}
         run: |
-          python -m pip install --upgrade pip
+          python -m pip install --upgrade pip setuptools_scm build
+          python -m setuptools_scm
           python -m pip install .[dev]
       -
         name: Run tests
@@ -62,7 +63,7 @@ jobs:
       - 
         name: Install dependencies
         run: |
-          python -m pip install --upgrade pip setuptools wheel setuptools_scm
+          python -m pip install --upgrade pip setuptools wheel setuptools_scm build
       - 
         name: Build package
         run: >-
