+++ setup.cfg
@@ -22,8 +22,6 @@ dev =
   torch
   tensorflow
   pytest
-  build
-  setuptools_scm
 
 [flake8]
 max-line-length = 88
