### 2022-08-09 b243547
make sure we fetch tag history for setuptools_scm
