+++ .github/workflows/run_tests_and_deploy.yml
@@ -23,6 +23,8 @@ jobs:
     steps:
       -
         uses: actions/checkout@v3
+        with:
+          fetch-depth: 0
       -
         name: Additional info about the build
         shell: bash
@@ -54,7 +56,9 @@ jobs:
     steps:
       - 
         name: Check out repository code
-        uses: actions/checkout@v2
+        uses: actions/checkout@v3
+        with:
+          fetch-depth: 0        
       - 
         name: Setup Python
         uses: actions/setup-python@v2
