### 2022-08-09 NREL/graph-env#34 @pstjohn
This was changed in #35 and pypi version 0.0.7; you can now separately install tensorflow or torch with
`pip install 'graphenv[tensorflow]'`
`pip install 'graphenv[torch]'`
while the base graphenv install only installs ray, networkx, and gym.

or install all the development dependencies (including pytest) with
`pip install 'graphenv[dev]'`
