### 2022-08-09 NREL/graph-env#33 @vwxyzjn
Thanks for entertaining with this. Yeah, it's a bit unfortunate that `ray` locking can take a long time. If it's too much pain, then https://github.com/NREL/graph-env/pull/35 looks good to me.

Feel free to close this.

As a side note — have you considered supporting other DRL libraries like Stable-baselines3?
