### 2022-08-10 #4621 @Viech
I have completed my review with some remarks below and will tick my remaining boxes when these points were addressed or discussed. Overall, I find that this is a small but nicely maintained and documented package whose scope could be described a bit better to a non-specialist audience.

## General checks

#### Substantial scholarly effort

Functionally this is a very small library: Excluding the `examples` subfolder, there are significantly less than 1000 lines of code in the `graphenv` folder that appears to contain the library as such, [so I should flag this](https://joss.readthedocs.io/en/latest/submitting.html) to @osorensen. This code gives the impression of a joint wrapper around two existing libraries. On the other hand, a substantial effort is apparent in the presentation and integration of this code in the form of tests, examples, and deploymet scripts. I can also see that the functionality offered is useful and not accessible to users not familiar with the backend libraries used.

## Functionality

#### Installation

- I have opened [#36](https://github.com/NREL/graph-env/issues/36).

## Documentation

#### A statement of need

- JOSS asks for the target audience to be mentioned in the documentation.

#### Installation instructions

- The dependencies are handled by `pip` but it would be good to have them listed also in the documentation (for working with the source and for external package managers).
- It appears that one can choose between `tensorflow` and `torch`, with one of them being required. This is not documented in the "Installation" section.

## Manuscript

I find that the manuscript is well-written and concise but the distinction between graph search problems and other problem representations could be made more clear as this seems to be the selling point for the software.

#### Author list

- If possible with the template, you should use inter-word spacing between "St." and "John".
- If possible, the digit 2 and the pilcrow in the author affiliations would look better on the respective next line, before their associated text. (Also, I would expect some contact info for the corresponding author; was this missed or is it not intended?)

#### Summary

- The concept of "strong relational inductive biases" is not made clear to a general audience.
- It would be nice to get an intuition on why certain problems are graph search representable but do not have an algebraic form (beyond mentioning molecular optimization as an example) as this seems to be the main motivation to use your package over classical optimization techniques.
- "or integer program" -> "or integer programs"?
- I don't understand why you mention "with well defined linear objective function and linear constraints": If linearity is important here, you could just write "integer linear programs". Otherwise, mentioning linearity does not support your argument that graph search problems are more expressive than classical representations.
- You use both a "traveling salesman problem" and "Traveling Salesman Problem" capitalization.
- You write "RLLib" while the associated paper writes it as "RLlib" in the title.

#### Statement of need

- JOSS asks for the target audience to be mentioned here.
- In the RLLib paragraph, it is not clear how the concepts of *parametrically defined* and *invalid actions*, *flattening*, *masking*, and *successor state* relate to what we learned in the summary. If these are central concepts, you should consider introducing them briefly in the summary. Otherwise this paragraph might be too technical for a statement of need.
