### 2022-08-10 #4621 @osorensen
Thanks for your review @Viech!

Regarding scholarly scope, the [guidelines](https://joss.readthedocs.io/en/latest/submitting.html) state as some factors to consider, which I will consider point-by-point below, while trying to "think out loud".

> - Age of software (is this a well-established software project) / length of commit history.

Looking at the repository, I find that the software seems rather young, with the [first commit](https://github.com/NREL/graph-env/commit/53916fde31b74c5a815c8677b53574b4cb51c9ec) made on February 11 2022.

> - Number of commits.

This is rather large, 241 in total.

> - Number of authors.

There have been four contributors, which suggests that this is a joint effort.

> - Total lines of code (LOC). Submissions under 1000 LOC will usually be flagged, those under 300 LOC will be desk rejected.

This one has already been commented on in @Viech's review.

> - Whether the software has already been cited in academic papers.

Could you please comment on this @pstjohn? Is the software used in academic paper, or are you aware of papers in preparation which will use the package.

> - Whether the software is sufficiently useful that it is likely to be cited by your peer group.

This is an important point. Making a joint wrapper around two existing libraries can in itself be very useful for users not familiar with those libraries.

In sum, I understand your concern about scholarly scope @Viech, and see that this is a difficult case. I would also be interested to hear what @iammix thinks of this, as the "Substantial scholarly effort" button has not been ticked off in your review checklist.

Also @Viech, could you think of any suggested extensions of the package that would make it more clearly within scope for JOSS?

You're of course also welcome to comment on this @pstjohn.
