### 2022-08-10 #4621 @idoby
@osorensen Just adding a comment here since I did check the respective checkbox. Following the discussion here I went back and checked the commit history in greater depth and found that a large number of commits are just one-liners or seem to be purely about cosmetics/experimenting with settings.

So I would suggest the authors

- document these successful (or less successful) configurations and experiments to convince the reader that their library is indeed robust and stands for more than just integration between various libraries (which is great but personally not something I would read a journal to find out about), and
- go back and clean up some of the history so that important development milestones stand out more clearly.

Personally, I believe git history to be more than just a storage space for old versions of the code, it also serves to document the design and development process for a project, document decisions made, etc.

(also, apologies, this is my first review and I didn't realize reviews could extend beyond the checklist)
