### 2022-08-11 #4621 @Viech
> Also @Viech, could you think of any suggested extensions of the package that would make it more clearly within scope for JOSS?

I'm more into "classical" optimization than machine learning, so I would not know about suited core features to be added. However one could certainly add more interface features: Default implementations of the `while not done` main loop with options for detailed output/logging of what is going on and control over stopping criteria, ability to stop, save, load, and continue a search, plotting of (parts of) the search graph, statistics concerning this graph, or additional ways to define a problem (e.g. by assigning callback functions to properties of some `GraphSearchProblem` instance so that the library can be used in an interactive console) are some potential features that come to my mind. I'm also just thinking out loud here; some of these ideas might not apply to `graphenv`. The general direction of these suggestions is that, ideally,

1. users of `graphenv` should be able to get basic functionality out of it without importing `gym` or `ray.rllib` themselves and/or
2. they should be offered advanced "quality of life" features if they are already comfortable using these backend libraries directly.

With either of these directions pursued, the package would be clearly within scope as I understand it.
