### 2022-08-15 NREL/graph-env#39 @davebiagioni
Update documentation
Addresses [JOSS review comments on documentation](https://github.com/openjournals/joss-reviews/issues/4621#issuecomment-1210934872)

- [ ] A statement of need. JOSS asks for the target audience to be mentioned in the documentation.

Installation instructions
- [ ] The dependencies are handled by pip but it would be good to have them listed also in the documentation (for working with the source and for external package managers).
- [ ] It appears that one can choose between tensorflow and torch, with one of them being required. This is not documented in the "Installation" section.
