### 2022-08-15 NREL/graph-env#40 @davebiagioni
Update manuscript
Addresses [JOSS review comments on manuscript](https://github.com/openjournals/joss-reviews/issues/4621#issuecomment-1210934872).

Author list
- [x] If possible with the template, you should use inter-word spacing between "St." and "John".
- [x] If possible, the digit 2 and the pilcrow in the author affiliations would look better on the respective next line, before their associated text.
    - **DB:  I cannot find an example of this, the example paper in the JOSS documentation doesn't follow this requirement.  I looked through the 10 most recent JOSS papers and found no examples there, either.  I believe this should not be required?**
- [x] Also, I would expect some contact info for the corresponding author; was this missed or is it not intended?
    - **Same comment as above.**

Summary
- [x] The concept of "strong relational inductive biases" is not made clear to a general audience.
- [x] It would be nice to get an intuition on why certain problems are graph search representable but do not have an algebraic form (beyond mentioning molecular optimization as an example) as this seems to be the main motivation to use your package over classical optimization techniques.
- [x] "or integer program" -> "or integer programs"?
- [x] I don't understand why you mention "with well defined linear objective function and linear constraints": If linearity is important here, you could just write "integer linear programs". Otherwise, mentioning linearity does not support your argument that graph search problems are more expressive than classical representations.
    - Thanks for catching this, we simply say "integer programs" now which is accurate.
- [x] You use both a "traveling salesman problem" and "Traveling Salesman Problem" capitalization.
- [x] You write "RLLib" while the associated paper writes it as "RLlib" in the title.

Statement of need
- [x] JOSS asks for the target audience to be mentioned here.
- [x] In the RLLib paragraph, it is not clear how the concepts of parametrically defined and invalid actions, flattening, masking, and successor state relate to what we learned in the summary. If these are central concepts, you should consider introducing them briefly in the summary. Otherwise this paragraph might be too technical for a statement of need.
    - We now mention these only parenthetically for user who might be familiar with these issues.
