### 2022-08-16 96dd254
adding copy, viz, and pickle examples to docs
