+++ _sources/_autosummary/graphenv.examples.tsp.tsp_state.TSPState.rst.txt
@@ -16,6 +16,7 @@ graphenv.examples.tsp.tsp\_state.TSPState
       :nosignatures:
    
       ~TSPState.new
+      ~TSPState.render
    
    
 
