+++ _sources/_autosummary/graphenv.vertex.Vertex.rst.txt
@@ -15,6 +15,7 @@ graphenv.vertex.Vertex
    .. autosummary::
       :nosignatures:
    
+      ~Vertex.render
    
    
 
