### 2022-08-16 NREL/graph-env#37 @pstjohn
I think our main goal with this library was to expose tree search as an environment under [the openai gym API](https://github.com/openai/gym), and enable it to be solved using ray's rllib algorithms. Here are some edits we've made in response to your suggestions, and where we think the existing implementations may be more sustainable:

> Default implementations of the `while not done` main loop

This is something we leave to the user, because the gym API is really centered around the call to env.step():
```python
observation, reward, done, info = env.step(action)
```
Adding 'default' implementations of this loop would essentially require us to to implement reinforcement learning policies, which we are intending to leave to third-party libraries (namely rllib).

> control over stopping criteria, ability to stop, save, load, and continue a search

We've added some [additional documentation](https://nrel.github.io/graph-env/examples/tsp_env.html#Beam-Search) on using graphenv for the TSP environment (without rllib) and for [saving/loading the environment](https://nrel.github.io/graph-env/examples/hallway.html#Saving-and-loading-GraphEnvs) in the Hallway example.

> plotting of (parts of) the search graph, statistics concerning this graph

The default `Vertex` class attempts to be as low-memory as possible and does not keep track of its history. Statistics are typically kept by the RL algorithm, and for instance are logged by rllib directly in a tensorboard instance. We have added a pass-through `env.render()` function to mimic the gym API, and demonstrated [constructing a movie of actions](https://nrel.github.io/graph-env/examples/tsp_env.html#) in the documentation.

>  additional ways to define a problem (e.g. by assigning callback functions to properties of some GraphSearchProblem instance so that the library can be used in an interactive console)

The way users interact with this library is by subclassing `Vertex` and defining their own search tree, observation, and reward functions. In our usage of the library, this has been more than enough flexibility without needing additional callbacks. If there was some functionality that needed an extra callback though, we'd be happy to add it.
