### 2022-08-16 e4b4199
add fetch depth for version on docs site
