+++ .github/workflows/build_docs.yml
@@ -5,7 +5,7 @@ on:
   # Triggers the workflow on push or pull request events
   push:
     tags:
-      - 'v*'      
+      - 'v*'
   #pull_request:
   #  branches: [ main ]
 
@@ -22,18 +22,20 @@ jobs:
         shell: bash -l {0}
     steps:
       - name: Checkout
-        uses: actions/checkout@v2
-        
+        uses: actions/checkout@v3
+        with:
+          fetch-depth: 0
+
       - name: Fetch all history for all tags and branches
         run: git fetch --prune --unshallow
-      
+
       - name: Configure Conda
         uses: conda-incubator/setup-miniconda@v2
         with:
           activate-environment: graphenv
           environment-file: devtools/conda-envs/environment.yml
           python-version: 3.9
-      
+
       - name: Build docs with Sphinx
         run: |
           cd docs
