### 2022-08-17 ef326b8
Better handling of setuptools_scm in sphinx documentation
Following documentation from https://github.com/pypa/setuptools_scm#usage-from-sphinx
