+++ .github/workflows/build_docs.yml
@@ -4,6 +4,8 @@ name: Create Docs
 on:
   # Triggers the workflow on push or pull request events
   push:
+    branches:
+      - 'main'
     tags:
       - 'v*'
   #pull_request:
