+++ .github/workflows/build_docs.yml
@@ -27,21 +27,18 @@ jobs:
         uses: actions/checkout@v3
         with:
           fetch-depth: 0
-
-      - name: Fetch all history for all tags and branches
-        run: git fetch --prune --unshallow
-
-      - name: Configure Conda
-        uses: conda-incubator/setup-miniconda@v2
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v4
         with:
-          activate-environment: graphenv
-          environment-file: devtools/conda-envs/environment.yml
-          python-version: 3.9
-
+          python-version: ${{ matrix.python-version }}
+      - name: Install package
+        shell: bash -l {0}
+        run: |
+          python -m pip install --upgrade pip setuptools_scm build
+          python -m pip install .[dev]
       - name: Build docs with Sphinx
         run: |
           cd docs
-          conda install pandoc
           pip install -r requirements_for_docs.txt
           #sphinx-apidoc -f -o source/ ../graphenv/
           make html
