+++ _sources/index.rst.txt
@@ -1,40 +1,0 @@
-.. graph-env documentation master file, created by
-   sphinx-quickstart on Fri Mar 18 11:43:00 2022.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-graph-env
-=========
-
-.. include:: background/introduction.rst
-
-Tutorials
-=========
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Hallway Example
-
-   examples/hallway
-   examples/tune
-
-.. toctree::
-   :maxdepth: 2
-   :caption: TSP Example
-
-   examples/tsp_docs
-   examples/tsp_env
-   cloud/tsp_on_aws
-
-.. toctree::
-   :hidden:
-   :caption: API reference
-
-   _autosummary/graphenv
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
