+++ _sources/_autosummary/graphenv.examples.tsp.tsp_nfp_model.TSPGNNQModel.rst.txt
@@ -1,40 +1,0 @@
-graphenv.examples.tsp.tsp\_nfp\_model.TSPGNNQModel
-==================================================
-
-.. currentmodule:: graphenv.examples.tsp.tsp_nfp_model
-
-.. autoclass:: TSPGNNQModel
-   :members:
-   :show-inheritance:
-   :special-members: __call__, __add__, __mul__
-
-   
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-      :nosignatures:
-   
-      ~TSPGNNQModel.context
-      ~TSPGNNQModel.custom_loss
-      ~TSPGNNQModel.forward
-      ~TSPGNNQModel.forward_vertex
-      ~TSPGNNQModel.from_batch
-      ~TSPGNNQModel.get_initial_state
-      ~TSPGNNQModel.get_q_value_distributions
-      ~TSPGNNQModel.get_state_value
-      ~TSPGNNQModel.import_from_h5
-      ~TSPGNNQModel.is_time_major
-      ~TSPGNNQModel.last_output
-      ~TSPGNNQModel.metrics
-      ~TSPGNNQModel.register_variables
-      ~TSPGNNQModel.trainable_variables
-      ~TSPGNNQModel.update_ops
-      ~TSPGNNQModel.value_function
-      ~TSPGNNQModel.variables
-   
-   
-
-   
-   
-   
