+++ _sources/examples/hallway.ipynb.txt
@@ -1,0 +1,675 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "id": "ddc3ca63",
+   "metadata": {},
+   "source": [
+    "# Hallway Environment Example"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "id": "a515e273",
+   "metadata": {},
+   "source": [
+    "The `graphenv` library implements graph search as a reinforcement learning (RL) problem with parameteric action spaces, and is ready to use with many off-the-shelf algorithms available in the RLLib framework.\n",
+    "\n",
+    "Before jumping into the implementation details, let's take a look a simple motivating example: the hallway problem.\n",
+    "\n",
+    "The hallway problem is effectively a 1d version of the gridworld problem in classic RL.  We are given a hallway with $N$ discrete positions and, starting at one end, want to learn to reach the opposite end in as few steps as possible.\n",
+    "\n",
+    "![hallway-flat](./img/hallway-flat.png)\n",
+    "\n",
+    "The figure above shows a hallway problem with $N=3$, and the optimal solution starting at state 0 and ending at state 2, with each \"current state\" highlighted in black.\n",
+    "\n",
+    "This trivial problem can be used to express the \"RL on a graph\" idea succinctly, and enable solving much more interesting, non-trivial problems."
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "id": "7cc44a24",
+   "metadata": {},
+   "source": [
+    "## The hallway problem as graph problem\n",
+    "\n",
+    "Before we jump into the graph formulation of the hallway problem, let's talk first about _actions_, because this is one of the key differences between `graphenv` and traditional RL gym environments.  \n",
+    "\n",
+    "Typically, discrete actions spaces in gym environments have fixed dimension and fixed definitions.  In the hallway problem, for instance, there are always two actions: \"move left\" (action=0) and \"move right\" (action=1).  Cases where the action is infeasible, like trying to move left from the start position, are typically handled by implmenting a _null_ transition where the action doesn't change the current state.\n",
+    "\n",
+    "In general graph search problems, however, such fixed action spaces are not practical.  In the game of chess, for example, the total number of possible board states and, subsequently, moves, is astronomical, while the set of _feasible_ moves changes continually throughout the game.\n",
+    "\n",
+    "The key observation that makes graph search tractable as a gym environment is:  even for large and complex graph search problems, the number of states that are accessible from the current state is usually relatively small.  If, instead of fixing the action to a pre-defined set of choices, we think of each action as representing an accessible next state, this endeavor becomes tractable.\n",
+    "\n",
+    "And so, we abandon the idea of \"fixed\" action spaces in favor of \"parametric\" action spaces.  Here, **parametric** means that actions coincide with next states represented by feature vectors, rather than having a single, index-based interpretation.  In other words, rather than actions as \"move $[ left, right ]$\", parametric actions can be thought of as \"go to state $[i, j]$\" where the states $i,j$ each have a vector representation.  Also unliked fixed spaces, the order doesn't matter:  we could equivalently say \"go to state $[j, i]$\".\n",
+    "\n",
+    "A key ingredient in making this machinery work is to have policy models that can work on parameteric action spaces\n",
+    "\n",
[…]
