+++ _sources/_autosummary/graphenv.graph_model.GraphModel.rst.txt
@@ -1,0 +1,26 @@
+graphenv.graph\_model.GraphModel
+================================
+
+.. currentmodule:: graphenv.graph_model
+
+.. autoclass:: GraphModel
+   :members:
+   :show-inheritance:
+   :special-members: __call__, __add__, __mul__
+
+   
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+      :nosignatures:
+   
+      ~GraphModel.forward
+      ~GraphModel.forward_vertex
+      ~GraphModel.value_function
+   
+   
+
+   
+   
+   

[…]
