### 2022-08-17 decdf78
addresses JOSS reviewer comments for the manuscript in #40
