### 2022-08-17 0579f32
Addresses JOSS reviewer comments for the manuscript (#44)
* addresses JOSS reviewer comments for the manuscript in #40

* Update introduction.rst

Co-authored-by: Biagioni <dbiagon@nrel.gov>
Co-authored-by: Peter St. John <peterc.stjohn@gmail.com>
