### 2022-08-17 #4621 @pstjohn
Hi all,

Thanks for the excellent comments (so far) on this effort! We've made a few changes to the repo and manuscript (see issues #33, #34, #36, #37, #39, and #40 for additional discussion), and here I'll respond generally to some of the bigger picture comments on the size, age, and git history that have been raised.

This library was originally conceived as a refactor to our [NREL/rlmolecule](github.com/nrel/rlmolecule) library, which used significantly more code to accomplish a similar task. We felt the amount of effort to integrate a tree search gym environment with a custom policy model in RLlib was nontrivial and might be of interest to the growing field of researchers looking at applying RL to new optimization problems. This in part explains the age of the library, our original work on these types of problems began in 2019, but we only recently refactored the code in this more general library. Our [recent publication](https://www.nature.com/articles/s42256-022-00506-3) cites that rlmolecule repository, although a lot of that work was ported here.

We subsequently spent a long time attempting to *reduce* the size of the library, trying to optimize speed and code maintainability. For instance, we removed the 177 line [space_util.py](https://github.com/NREL/graph-env/blob/ef60e5fb9b9111c89cdf304187aec5dba38cd6bd/graphenv/space_util.py) functions after finding ways to leverage the built-in `repeated` action space within rllib.

This also explains in part the large number of "small" commits; debugging RLlib can be challenging (as most of the work occurs in forked worker processes) and as a result we took a rather circuitous route to our current implementation. We're documenting current "gotcha's" in our online documentation, but could add a "Lessons Learned" section to our documentation moving forward to keep track of knowledge gained through failed experiments.

On re-writing the git history, I'm not confident enough in my git abilities to force-push `main` on a repo with PyPI integration and automated docs tied directly to SCM version tags :cold_sweat:. But moving forward, we will certainly package major development milestones into pull requests and merge those as a single commit.

We appreciate everyone's involvement so far! Please let us know if any of our responses could use additional detail. We put a response together around project scope in #37, and would be happy to continue those discussions there or in this thread.

(signed all authors)
