### 2022-08-18 #4621 @Viech
I've reviewed the changes to the manuscript and documentation and they look good to me!

With respect to scope I want to stress that I can see this package being used and cited in research and that the code and repository quality look pretty good; so I've ticked that box now even though not much new substance was added to the core code at this point. I also understand that using compact constructs can improve code quality while reducing size and that the efficient use of a technical backend library can be a feature in itself.

If I may chime in concerning a history rewrite: I think this should not be relevant for publication. I see the value that a nice history has for understanding the ideas behind pieces of code but rewriting git history is typically considered bad style as it breaks forks, branches, and leaves important tags detached from the new history. I would suggest this in extreme cases but not here.
