### 2022-08-24 dd41534
Bump nbconvert from 6.5.0 to 6.5.1 in /docs (#45)
Bumps [nbconvert](https://github.com/jupyter/nbconvert) from 6.5.0 to 6.5.1.
- [Release notes](https://github.com/jupyter/nbconvert/releases)
- [Commits](https://github.com/jupyter/nbconvert/compare/6.5...6.5.1)

---
updated-dependencies:
- dependency-name: nbconvert
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: Peter St. John <peterc.stjohn@gmail
[…]
