+++ docs/requirements_for_docs.txt
@@ -1,5 +1,5 @@
 ipython==8.2.0
-nbconvert==6.5.0
+nbconvert==6.5.1
 nbsphinx==0.8.8
 Sphinx==4.5.0
 sphinx-rtd-theme==1.0.0
