### 2022-08-27 9f2ce9a
UPDATE: Fix typos, PEP8 python format, correct declaration of return type
