### 2022-08-27 #4621 @iammix
Thanks @pstjohn for your submission. First of all, you have done remarkable work on it. Nowadays, it is generally accepted to combine different tools to create a wrapper of them. In my opinion, this will be the future of research.
You have a well-organized documentation page, where you have provided the background of the examples. "Lesson Learned" is a great add-on to the documentation and I am looking forward to that.

As far as the code is concerned, I noticed that it is easy for someone to be able to extend the code and create their custom methods(new features etc). Moreover, you impemented the relatively new feature of python(typing) where you declare the type of arguments and returns of your methods.

I've done some minor changes(fix typos, return types at the declaration of the methods, [PEP8 python Style Guide format](https://peps.python.org/pep-0008/) ). Check my PR https://github.com/NREL/graph-env/pull/47
