### 2022-08-27 5a4ce42
UPDATE: Fix typos, PEP8 python format, correct declaration of return type (#47)
