### 2022-08-27 #4621 @pstjohn
Thanks @iammix for the detailed review and fixes!! I've merged https://github.com/NREL/graph-env/pull/47, I'll pin a new release (likely 0.1.0) if/when the paper is accepted.

@osorensen, let us know what the next steps would be

Thanks again everyone for the helpful comments
