### 2022-08-27 #4621 @osorensen
@pstjohn At this point could you:

In the meantime I'll read through the manuscript again and let you know if I have any suggested changes.
