### 2022-08-28 NREL/graph-env#48 @osorensen
JOSS submission
I'll open issues for my editorial comments on the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4621).

The first one is this. Please remove the extra linespace at line 11, after "inductive biases":
