### 2022-08-28 NREL/graph-env#49 @osorensen
JOSS submission
Another editorial comments on the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4621)

On lines 15 and 30 you define the abbreviation TSP

Then on line 41 you write "Traveling Salesperson Problem":

I suggest defining "Travling Salesperson Problem (TSP)" on line 15, and the writing simply "TSP" on lines 30 and 41.
