### 2022-08-28 NREL/graph-env#50 @osorensen
DOI for arXiv preprints
Another editorial comment on https://github.com/openjournals/joss-reviews/issues/4621.

arXiv now has DOIs for its preprints. Please update the `.bib` file accordingly, so the DOIs show up for all arXiv references. Here is one example: https://arxiv.org/abs/1806.01261

The BibTeX entry for this paper should be something like:

```
@misc{https://doi.org/10.48550/arxiv.1806.01261,
  doi = {10.48550/ARXIV.1806.01261},

  url = {https://arxiv.org/abs/1806.01261},

  author = {Battaglia, Peter W. and Hamrick, Jessica B. and Bapst, Victor and Sanchez-Gonzalez, Alvaro and Zambaldi, Vinicius and Malinowski, Mateusz and Tacchetti, Andrea and Raposo, David and Santoro, Adam and Faulkner, Ryan and Gulcehre, Caglar and Song, Francis and Ballard, Andrew and Gilmer, Justin and Dahl, George and Vaswani, Ashish and Allen, Kelsey and Nash, Charles and Langston, Victoria and Dyer, Chris and Heess, Nicolas and Wierstra, Daan and Kohli, Pushmeet and Botvinick, Matt and Vinyals, Oriol and Li, Yujia and Pascanu, Razvan},

  keywords = {Machine Learning (cs.LG), Artificial Intelligence (cs.AI), Machine Learning (stat.ML), FOS: Computer and information sciences, FOS: Computer and information sciences},

  title = {Relational inductive biases, deep learning, and graph networks},

  publisher = {arXiv},

  year = {2018},

  copyright = {arXiv.org perpetual, non-exclusive license}
}

```
