+++ joss/paper.bib
@@ -7,7 +7,9 @@
 		  and Malinowski, Mateusz and Tacchetti, Andrea and Raposo,
 		  David and Santoro, Adam and Faulkner, Ryan and others},
   year		= 2018,
-  journal	= {arXiv preprint arXiv:1806.01261}
+  journal	= {arXiv preprint arXiv:1806.01261},
+  doi = {10.48550/arXiv.1806.01261},
+  url = {https://arxiv.org/abs/1806.01261}
 }
 
 @TechReport{	  biagioni2020rlmolecule,
@@ -19,7 +21,9 @@
 		  others},
   year		= 2020,
   institution	= {National Renewable Energy Lab.(NREL), Golden, CO (United
-		  States)}
+		  States)},
+  doi = {10.11578/dc.20201221.3},
+  url = {https://www.osti.gov/biblio/1737555}
 }
 
 @Article{	  brockman2016openai,
@@ -28,7 +32,9 @@
 		  and Schneider, Jonas and Schulman, John and Tang, Jie and
 		  Zaremba, Wojciech},
   year		= 2016,
-  journal	= {arXiv preprint arXiv:1606.01540}
+  journal	= {arXiv preprint arXiv:1606.01540},
+  url = {https://arxiv.org/abs/1606.01540},
+  doi = {10.48550/arXiv.1606.01540}
 }
 
 @InProceedings{	  liang2018rllib,
@@ -40,7 +46,9 @@
   year		= 2018,
   booktitle	= {International Conference on Machine Learning},
   pages		= {3053--3062},
-  organization	= {PMLR}
+  organization	= {PMLR},
+  url = {https://arxiv.org/abs/1712.09381},
+  doi = {10.48550/arXiv.1712.09381}
 }
 
 @Article{	  Pandey_2021,
@@ -102,13 +110,16 @@
 }
 
 @Article{	  sv2021multi,
-  title		= {Multi-objective goal-directed optimization of de novo stable organic radicals for aqueous redox flow batteries},
-  author	= {S.V., Shree S. and Law, Jeffrey and Tripp, Charles
-		  and Duplyakin, Dmitry and Skordilis, Erotokritos and
-		  Biagioni, David and Paton, Robert and {St. John}, Peter},
+  title		= {Multi-objective goal-directed optimization of de novo
+		  stable organic radicals for aqueous redox flow batteries},
+  author	= {Shree Sowndarya S. V. and Jeffrey N. Law and Charles E.
+		  Tripp and Dmitry Duplyakin and Erotokritos Skordilis and
+		  David Biagioni and Robert S. Paton and Peter C. St. John},
   year		= 2022,
-  journal = {Nature Machine Intelligence},
-  doi = {10.1038/s42256-022-00506-3}
+  month		= {aug},
+  journal	= {Nature Machine Intelligence},
+  publisher	= {Springer Science and Business Media {LLC}},
+  doi		= {10.1038/s42256-022-00506-3}
 }
 
 @InCollection{	  Zheng_2020,
@@ -128,7 +139,8 @@
     author={Antoine Prouvost and Justin Dumouchelle and Lara Scavuzzo and Maxime Gasse and Didier Ch{\'e}telat and Andrea Lodi},
     booktitle={Learning Meets Combinatorial Algorithms at NeurIPS2020},
     year={2020},
-    url={https://openreview.net/forum?id=IVc9hqgibyB}
+    url={https://openreview.net/forum?id=IVc9hqgibyB},
+    doi={10.48550/arXiv.2011.06069}
 }
 
 @Article{	  Zhou_2019,
