### 2022-08-28 0a3aefd
Joss paper edits (#51)
* resolve #50

* delete comments and resolve #48 , #49

* Create .zenodo.json
