+++ joss/paper.md
@@ -1,18 +1,12 @@
 ---
-title: 'graphenv: a Python library for reinforcement learning on graph search spaces'
+title:
+  'graphenv: a Python library for reinforcement learning on graph search spaces'
 tags:
   - Python
   - reinforcement learning
   - graph search
   - combinatorial optimization
 authors:
-  # - name: Adrian M. Price-Whelan^[Co-first author] # note this makes a footnote saying 'Co-first author'
-  #   orcid: 0000-0000-0000-0000
-  #   affiliation: "1, 2" # (Multiple affiliations must be quoted)
-  # - name: Author Without ORCID^[Co-first author] # note this makes a footnote saying 'Co-first author'
-  #   affiliation: 2
-  # - name: Author with no affiliation^[Corresponding author]
-  #   affiliation: 3
   - name: David Biagioni
     orcid: 0000-0001-6140-1957
     affiliation: 2
@@ -21,146 +15,119 @@ authors:
     affiliation: 2
   - name: Struan Clark
     affiliation: 2
-    orcid:  0000-0003-0078-6560
+    orcid: 0000-0003-0078-6560
   - name: Dmitry Duplyakin
     affiliation: 2
-    orcid:  0000-0001-5132-0168
+    orcid: 0000-0001-5132-0168
   - name: Jeffrey Law
     affiliation: 1
-    orcid: 0000-0003-2828-1273    
-  - name: Peter C. St. John #^[Corresponding author]
+    orcid: 0000-0003-2828-1273
+  - name: Peter C. St. John
     orcid: 0000-0002-7928-3722
-    corresponding: true # (This is how to denote the corresponding author)    
-    affiliation: 1    
+    corresponding: true
+    affiliation: 1
 affiliations:
- - name: Biosciences Center, National Renewable Energy Laboratory, Golden CO 80401, USA
-   index: 1
- - name: Computational Sciences Center, National Renewable Energy Laboratory, Golden CO 80401, USA
-   index: 2
-
-#  - name: Institution Name, Country
-#    index: 2
-#  - name: Independent Researcher, Country
-#    index: 3
+  - name:
+      Biosciences Center, National Renewable Energy Laboratory, Golden CO 80401,
+      USA
+    index: 1
+  - name:
+      Computational Sciences Center, National Renewable Energy Laboratory,
+      Golden CO 80401, USA
+    index: 2
 date: 17 May 2022
 bibliography: paper.bib
-
-# Optional fields if submitting to a AAS journal too, see this blog post:
-# https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
-# aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
-# aas-journal: Astrophysical Journal <- The name of the AAS journal.
 ---
 
 # Summary
 
-Many important and challenging problems in combinatorial optimization (CO) can be
-expressed as graph search problems, in which graph vertices represent full or partial
-solutions and edges represent decisions that connect them. 
-Graph structure not only introduces strong _relational inductive biases_  
-for learning [@battaglia2018relational] -- in this context, by providing a way to 
-explicitly model the value of transitioning (along edges) between one search state (vertex) 
-and the next -- but lends itself to problems both with and without clearly 
-defined algebraic structure.  For example, classic CO problems on graphs such as the Traveling Salesman Problem (TSP) can be expressed as either pure graph search _or_ integer programs.  Other problems, however, such as molecular optimization, do no have concise algebraic formulations and yet are readily
-implemented as a graph search [@Zhou_2019;@sv2021multi].  Such "model-free" problems constitute
[…]
