### 2022-08-28 #4621 @pstjohn
Took a couple tries to get the .zenodo.json metadata in the right format, but I think I've made the requested paper edits and these associated tags:

`v0.1.3`
[10.5281/zenodo.7030161](https://doi.org/10.5281/zenodo.7030161)
